# rest/config/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Liverpool FC API",
    "version": "1.0.0",
    "description": "REST API for managing Liverpool Football Club data including players, matches, and more"
  },
  "servers": [
    {
      "url": "http://localhost:8989/api/v1"
    }
  ],
  "paths": {
...
    "/stadiums": {
      "get": {
        "summary": "Get all stadiums",
        "tags": [
          "Stadiums"
        ],
        "responses": {
          "200": {
            "description": "List of stadiums",
            "headers": {
              "X-Total-Count": {
                "description": "Total number of stadiums",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Stadium"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No stadiums found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create new stadium",
        "tags": [
          "Stadiums"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Stadium"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stadium created",
            "headers": {
              "Location": {
                "description": "URL of created stadium",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/StadiumResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Stadium with this name already exists"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/stadiums/{id}": {
      "get": {
        "summary": "Get stadium by ID",
        "tags": [
          "Stadiums"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stadium ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Stadium found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/StadiumResponse"
                }
              }
            }
          },
          "404": {
            "description": "Stadium not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update stadium",
        "tags": [
          "Stadiums"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stadium ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Stadium"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stadium updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/StadiumResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Stadium not found"
          },
          "409": {
            "description": "Stadium with this name already exists"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Partially update stadium",
        "tags": [
          "Stadiums"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stadium ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "capacity": {
                    "type": "number",
                    "minimum": 100
                  },
                  "location": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stadium updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/StadiumResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "404": {
            "description": "Stadium not found"
          },
          "409": {
            "description": "Stadium with this name already exists"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete stadium",
        "tags": [
          "Stadiums"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Stadium ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Stadium deleted"
          },
          "404": {
            "description": "Stadium not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Stadium": {
        "type": "object",
        "required": [
          "name",
          "capacity",
          "location"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Stadium name (must be unique)"
          },
          "capacity": {
            "type": "number",
            "minimum": 100,
            "description": "Stadium capacity (minimum 100 seats)"
          },
          "location": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Stadium location"
          }
        }
      },
...
    },
    "responses": {
...
      "StadiumResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Stadium"
          },
          "_links": {
            "type": "object",
            "properties": {
              "self": {
                "type": "string"
              },
              "collection": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
}